Update tests after refactor
