+++ karma.conf.js
@@ -63,6 +63,15 @@ module.exports = function (config) {
             'src/components/voteUrlProcessor/index.js',
             'src/components/onboarding/index.js',
             'src/components/passphrase/createSecond/index.js',
+            'src/components/passphraseCreation/index.js',
+            'src/components/registerDelegate/index.js',
+            'src/components/registerDelegate/registerDelegate.js',
+            'src/components/registerDelegate/steps/choose/choose.js',
+            'src/components/registerDelegate/steps/choose/index.js',
+            'src/components/registerDelegate/steps/confirm/confirm.js',
+            'src/components/registerDelegate/steps/confirm/index.js',
+            'src/store/reducers/delegate.js',
+            'src/actions/delegate.js',
             'src/utils/api/liskService.js',
           ],
           overrides: {
